# client/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ln-server",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/callbacks/open-channel": {
      "get": {
        "tags": ["ln-gateway"],
        "operationId": "openChannel",
        "parameters": [
          {
            "name": "k1",
            "in": "query",
            "description": "One-time token from /channel-request",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "remote_id",
            "in": "query",
            "description": "Remote node pubkey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "amount",
            "in": "query",
            "description": "Channel funding amount in satoshis",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "announce",
            "in": "query",
            "description": "Whether to announce channel",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Open channel result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenChannelResponse"
                }
              }
            }
          }
        }
      }
    },
    "/callbacks/withdraw-request": {
      "get": {
        "tags": ["ln-gateway"],
        "operationId": "withdraw",
        "parameters": [
          {
            "name": "k1",
            "in": "query",
            "description": "One-time token from /withdraw-request",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "destination",
            "in": "query",
            "description": "Bitcoin address (or other supported withdraw destination)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "amount",
            "in": "query",
            "description": "Withdraw amount in satoshis",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Withdraw result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueWithdrawResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IssueWithdrawResponse": {
        "type": "object",
        "required": ["tx", "psbt", "txid"],
        "properties": {
          "psbt": {
            "type": "string"
          },
          "tx": {
            "type": "string"
          },
          "txid": {
            "type": "string"
          }
        }
      },
...
      "OpenChannelResponse": {
        "type": "object",
        "required": ["ok"],
        "properties": {
          "error": {
            "type": ["string", "null"]
          },
          "ok": {
            "type": "boolean"
          },
          "result": {}
        }
      },
...
    }
  },
...
}